cleaning up code and making it simpler
