+++ .gitignore
@@ -23,3 +23,5 @@ backlinks.txt
 compare.md
 
 old/
+.git
+
