+++ CHANGELOG
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.7.0] - 2024-07-11
+### Cleanup
+- Removed the server
+- Removed the Dockerfile
+
 ## [0.6.1] - 2019-10-03
 ### Improvements
 - Cleaned up the help output
 - Added interface for an experimental storage backend
-
 
 ## [0.6.0] - 2019-07-30
 ### Improvements
