--- app.yaml
@@ -1,10 +1,0 @@
-service: urlinsane-api
-
-runtime: custom
-env: flex
-
-manual_scaling:
-  instances: 1
-
-network:
-  session_affinity: true
