--- cloudbuild.deploy.yaml
@@ -1,20 +1,0 @@
-steps:
-- name: 'gcr.io/cloud-builders/gcloud'
-  id: Push
-  args:
-  - 'builds'
-  - 'submit'
-  - '--tag'
-  - 'gcr.io/$PROJECT_ID/urlinsane'
-  waitFor:
-  - '-'
-- name: 'gcr.io/cloud-builders/gcloud'
-  id: Deploy
-  args:
-  - 'app'
-  - 'deploy'
-  - '--quiet'
-  - '--image-url'
-  - 'gcr.io/$PROJECT_ID/urlinsane'
-  waitFor:
-  - 'Push'
