+++ cmd/urlinsane/root-old.go
@@ -1,0 +1,79 @@
+// // Copyright (C) 2024  Rangertaha <rangertaha@gmail.com>
+// //
+// // This program is free software: you can redistribute it and/or modify
+// // it under the terms of the GNU General Public License as published by
+// // the Free Software Foundation, either version 3 of the License, or
+// // (at your option) any later version.
+// //
+// // This program is distributed in the hope that it will be useful,
+// // but WITHOUT ANY WARRANTY; without even the implied warranty of
+// // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// // GNU General Public License for more details.
+// //
+// // You should have received a copy of the GNU General Public License
+// // along with this program.  If not, see <http://www.gnu.org/licenses/>.
+package cmd
+
+// import (
+// 	// "bytes"
+// 	// "fmt"
+// 	"os"
+
+// 	// homedir "github.com/mitchellh/go-homedir"
+// 	"github.com/spf13/cobra"
+// 	// "github.com/spf13/viper"
+// )
+
+// // var cfgFile string
+
+// // const cfgFileContent = `
+
+// // storage:
+// // - elastic
+
+// // elastic:
+// //   host: 127.0.0.1
+// //   port: 32772
+// //   index: urlinsane
+
+// // typos:
+// // - all
+
+// // info:
+// // - all
+
+// // `
+
+// // // rootCmd represents the base command when called without any subcommands
+// // var rootCmd = &cobra.Command{
+// // 	Use:   "urlinsane",
+// // 	Short: "Generates domain typos and variations",
+// // 	Long: `
+// // Multilingual domain typo permutation engine used to perform or detect typosquatting, brandjacking,
+// // URL hijacking, fraud, phishing attacks, corporate espionage and threat intelligence.`,
+// // 	Run: func(cmd *cobra.Command, args []string) {
+// // 		if len(args) == 0 {
+// // 			cmd.Help()
+// // 			os.Exit(0)
+// // 		}
+
+// // 	},
+// // }
+
+// // Execute adds all child commands to the root command and sets flags appropriately.
+// // This is called by main.main(). It only needs to happen once to the rootCmd.
+// func Execute() {
+// 	if err := rootCmd.Execute(); err != nil {
+// 		// fmt.Println(err)
+// 		// os.Exit(1)
+// 	}
+// }
+
+// func init() {
+// 	// cobra.OnInitialize(initConfig)
+// 	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "Configuration file (default is $HOME/.urlinsane.yaml)")
+
+// 	// viper.SetDefault("elastic.port", "9200")
+// 	// viper.SetDefault("elastic.host", "127.0.0.1")
+// 	// viper.SetDefault("elastic.index", "urlinsane")
+// }
