+++ cmd/urlinsane/server.go
@@ -1,47 +1,41 @@
-// Copyright © 2018 Rangertaha <rangertaha@gmail.com>
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
+// // Copyright (C) 2024  Rangertaha <rangertaha@gmail.com>
+// //
+// // This program is free software: you can redistribute it and/or modify
+// // it under the terms of the GNU General Public License as published by
+// // the Free Software Foundation, either version 3 of the License, or
+// // (at your option) any later version.
+// //
+// // This program is distributed in the hope that it will be useful,
+// // but WITHOUT ANY WARRANTY; without even the implied warranty of
+// // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// // GNU General Public License for more details.
+// //
+// // You should have received a copy of the GNU General Public License
+// // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 package cmd
 
-import (
-	"github.com/spf13/cobra"
+// import (
+// 	"github.com/spf13/cobra"
 
-	"github.com/cybint/urlinsane/server"
-)
+// 	"github.com/cybint/urlinsane/server"
+// )
 
-// serverCmd represents the server command
-var serverCmd = &cobra.Command{
-	Use:   "server",
-	Short: "Start a websocket server to use this tool programmatically",
-	Long:  `This command starts up a REST API server to use this tool programmatically.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		host, _ := cmd.Flags().GetString("host")
-		port, _ := cmd.Flags().GetString("port")
-		concurrency, _ := cmd.Flags().GetInt("concurrency")
-		server.NewWebSocketServer(host, port, concurrency)
-	},
-}
+// // serverCmd represents the server command
+// var serverCmd = &cobra.Command{
+// 	Use:   "server",
+// 	Short: "Start a websocket server to use this tool programmatically",
+// 	Long:  `This command starts up a REST API server to use this tool programmatically.`,
+// 	Run: func(cmd *cobra.Command, args []string) {
+// 		host, _ := cmd.Flags().GetString("host")
+// 		port, _ := cmd.Flags().GetString("port")
+// 		concurrency, _ := cmd.Flags().GetInt("concurrency")
+// 		server.NewWebSocketServer(host, port, concurrency)
+// 	},
+// }
 
-func init() {
-	rootCmd.AddCommand(serverCmd)
-	serverCmd.Flags().StringP("host", "a", "127.0.0.1", "IP address for API server")
-	serverCmd.Flags().StringP("port", "p", "8080", "Port to use")
-	serverCmd.Flags().IntP("concurrency", "c", 50, "Number of concurrent workers")
-}
+// func init() {
+// 	rootCmd.AddCommand(serverCmd)
+// 	serverCmd.Flags().StringP("host", "a", "127.0.0.1", "IP address for API server")
+// 	serverCmd.Flags().StringP("port", "p", "8080", "Port to use")
+// 	serverCmd.Flags().IntP("concurrency", "c", 50, "Number of concurrent workers")
+// }
