+++ typo/info.go
