+++ typo/languages/arabic.go
