+++ typo/languages/armenian.go
