+++ typo/languages/persian.go
