+++ typo/languages/spanish.go
